Clean up MANIFEST before creating a new one.
This is a follow-up to 9899ff7ca05, which changed behavior from
appending to an existing MANIFEST to creating a new one.  On VMS,
that would give us multiple versions, which probably wouldn't
matter except that the tests insist that there is one and only
one copy of every file they've created and they only clean up the
highest version.

The easiest way to handle this is to make sure we clean up first
before creating a new MANIFEST file.