save full path to default MANIFEST.SKIP
The default MANIFEST.SKIP file is in the same directory as the
Manifest.pm file. EUM finds it by using __FILE__. __FILE__ will be a
relative path if it is loaded by a relative @INC path. This means that
if the current directory is changed, MANIFEST.SKIP won't be able to be
found.

Calculate at absolute path to MANIFEST.SKIP at module load time, where
__FILE__ is guaranteed to valid.