Skip iCloud placeholder files
macOS's iCloud has a feature to "optimize Mac storage" that replaces the real file with a placeholder file. It looks the same in the Finder but is really this file with .icloud added. When you click on it, it downloads the actual file. So, it's probably smart to not have repos in iCloud, but here we are.